+++ Leomard/Services/PersonService.swift
@@ -14,8 +14,8 @@ class PersonService: Service {
         self.requestHandler = requestHandler
     }
     
-    public func getPersonDetails(person: Person, page: Int, savedOnly: Bool, completion: @escaping (Result<GetPersonDetailsResponse, Error>) -> Void) {
-        var request = "/user?person_id=\(String(person.id))&page=\(String(page))"
+    public func getPersonDetails(person: Person, page: Int, savedOnly: Bool, sortType: SortType, completion: @escaping (Result<GetPersonDetailsResponse, Error>) -> Void) {
+        var request = "/user?person_id=\(String(person.id))&page=\(String(page))&sort=\(String(describing: sortType))"
         if savedOnly {
             request += "&saved_only=true"
         }