+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 - Added status indicator when sending a post (#7)
 - Added alert when sending/editing post/comment fails
 - Refreh button in the inbox
+- Replies sort method for inbox
+- Profile view sort method
 
 ### Bug Fixes